refactor(DirectionUtil) : util 패키지가 아닌 domain 패키지로 이동
[관련 피드백 링크](https://github.com/woowacourse/java-chess/pull/721#discussion_r1535530149)
